backend/requirements.txt
django==4.2
python-dotenv==1.1.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-extra-fields==3.7.0
djoser==2.3.1
pillow==11.2.1
hashids==1.3.1
